refactor: isolate Link component from wallet
